Fix to include GLO SVs with valid Eph or Alm data.
Fix to include GLO SVs with valid Ephemeris or Almanac data to be
included for SVs used in fix. The flag was set to zero because modem
GLO Eph is empty even though it had valid Alm data.

Also in addition to check for availability of Ephermeris or Alamanac
data, also check for SNR > 0 to indicate SV is used in fix. The
reason is even if modem stops tracking some SV’s, it reports them
in the measurement report with an Eph/Alm data with 0 SNR.

Change-Id: I67fb29e1937bd9d8ad5e68129feffb9b275577ef
CRs-Fixed: 1034666